#660 - store - split order, payment, product modules
