+++ src/store/modules/order/actions.js
@@ -1,0 +1,24 @@
+import * as types from 'src/store/mutation-types'
+import EventBus from 'src/plugins/event-bus'
+import { ValidationError } from 'lib/exceptions'
+const Ajv = require('ajv') // json validator
+
+export default {
+  /**
+   * Place order - send it to service worker queue
+   * @param {Object} commit method
+   * @param {Object} order order data to be send
+   */
+  placeOrder ({ commit }, order) {
+    const ajv = new Ajv()
+    const validate = ajv.compile(require('../../../models/order_schema.json'))
+
+    if (!validate(order)) { // schema validation of upcoming order
+      throw new ValidationError(validate.errors)
+    } else {
+      commit(types.ORDER_PLACE_ORDER, order)
+      EventBus.$emit('order-after-placed', { order: order })
+      return true
+    }
+  }
+}
