--- src/store/modules/payment.js
@@ -1,25 +1,0 @@
-
-import paymentMethods from '../../resource/payment_methods.json'
-
-const state = {
-  methods: paymentMethods
-}
-
-const getters = {
-}
-
-// actions
-const actions = {
-}
-
-// mutations
-const mutations = {
-}
-
-export default {
-  namespaced: true,
-  state,
-  getters,
-  actions,
-  mutations
-}
