+++ src/store/modules/payment/index.js
@@ -1,0 +1,8 @@
+import paymentMethods from 'src/resource/payment_methods.json'
+
+export default {
+  namespaced: true,
+  state: {
+    methods: paymentMethods
+  }
+}
