+++ src/store/modules/product/getters.js
@@ -1,0 +1,8 @@
+export default {
+  productParent: (state) => state.parent,
+  productCurrent: (state) => state.current,
+  currentConfiguration: (state) => state.current_configuration,
+  productOriginal: (state) => state.original,
+  currentOptions: (state) => state.current_options,
+  breadcrumbs: (state) => state.breadcrumbs
+}
